moved arr.h into utils.h
