+++ CppSundial/src/base/baseImpl.cpp
@@ -3,7 +3,6 @@
 #include "str.h"
 #include "hashtable.h"
 #include "allocators.h"
-#include "arr.h"
 #include "utils.h"
